Address PR feedback items
